Fix CachingHiveMetastore.getPartitionsByNames failing while cache invalidation
`CachingHiveMetastore` uses `HivePartitionName` for keys in partition
cache.  `HivePartitionName` has a peculiar, semi- value-based equality.
HivePartitionName may or may not be missing a name and it matters for
bulk load, but it doesn't matter for a single-partition load. Because
of equality semantics, the cache keys may gets mixed during bulk load.
This happens within `EvictableCache.getAll` where we compute tokens. We
may get an existing token and normally it's fine, since existing entry
is satisfied without a bulk load. However, if existing token has no data
in the `EvictableCache.dataCache` (e.g. due to explicit or implicit
invalidation), the loading gets invoked. Thus, a bulk loading may get a
`HivePartitionName` without a name, which is unexpected (and
unsupported) in the bulk loading.